requirements.txt
# Voice Interview Agent Dependencies

# Core
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0

# Groq SDK (STT + LLM)
groq>=0.9.0

# Deepgram SDK (TTS)
deepgram-sdk>=3.5.0

# Audio Processing
sounddevice>=0.4.6
numpy>=1.26.0
scipy>=1.12.0
pydub>=0.25.1

# Web Server (optional for API)
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Utilities
rich>=13.7.0
